Started implementing state pattern
Added a State interface with state classes so we could implement the state pattern and get better control over the games. We might have to make Whack-A-Mole extend Game instead of FreeDraw for this to work. The structure is there so we should just have to reformat the Game classes to use Game's new abstract methods.